refactor: moved tests to __tests__
