+++ __mocks__/cowsay.ts
