+++ __tests__/components/cowsay/Cowsay.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { CowsayTypes } from '@ducks/cowsay/types';
-import { connectedRender } from '../../tests/testUtils';
-import Cowsay from './Cowsay';
+import { connectedRender } from '@tests/testUtils';
+import Cowsay from '@components/cowsay/Cowsay';
 
 const cowsayState = {
   cowsay: {
