+++ __tests__/components/landingPage/LandingPage.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import * as keycloakpackage from '@react-keycloak/web';
 
-import * as keycloakMock from '@services/mocks';
+import * as keycloakMock from '@mocks/services';
+import * as mocks from '@mocks/landingPage';
 
-import { connectedRender } from '../../tests/testUtils';
-import * as mocks from './mocks';
-import LandingPage from './LandingPage';
+import { connectedRender } from '@tests/testUtils';
+import LandingPage from '@components/landingPage/LandingPage';
 
 // let's mock it another way
 jest.mock('@react-keycloak/web');
