+++ __tests__/components/mainScreen/MainScreen.test.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
 import * as keycloakpackage from '@react-keycloak/web';
 
-import * as keycloakMock from '@services/mocks';
+import * as keycloakMock from '@mocks/services';
 
-import { connectedRender, fireEvent } from '../../tests/testUtils';
-import MainScreen from './MainScreen';
+import { connectedRender, fireEvent } from '@tests/testUtils';
+import MainScreen from '@components/mainScreen/MainScreen';
 
-jest.mock('../cowsay/Cowsay', () => () => 'cowsay-mock');
-jest.mock('../landingPage/LandingPage', () => () => 'landing-page-mock');
+jest.mock('@components/cowsay/Cowsay', () => () => 'cowsay-mock');
+jest.mock(
+  '@components/landingPage/LandingPage',
+  () => () => 'landing-page-mock'
+);
 
 jest.mock('@react-keycloak/web');
 const useKeycloakMock = keycloakpackage as jest.Mocked<any>;
