+++ __tests__/store/ducks/cowsay/sagas.test.ts
@@ -2,9 +2,9 @@ import { expectSaga } from 'redux-saga-test-plan';
 import { throwError } from 'redux-saga-test-plan/providers';
 import * as matchers from 'redux-saga-test-plan/matchers';
 import api from '@services/api';
-import load from './sagas';
-import * as actions from './actions';
-import mocks from './mocks';
+import load from '@ducks/cowsay/sagas';
+import * as actions from '@ducks/cowsay/actions';
+import mocks from '@mocks/cowsay';
 
 describe('cowsay saga', () => {
   test('fetches the cowsay data', () => {
